% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crePalette.R
\name{crePalette}
\alias{crePalette}
\title{crePalette}
\usage{
crePalette(cols)
}
\arguments{
\item{cols}{single column data frame labeled HEX in which the hexadecimal color codes have been stored}
}
\value{
a vector of color strings directly associated with the hexadecimal values inputted
}
\description{
this function takes in a user inputted string of hexadecimal colors and converts them to a vector of color names to aid in data visualization
}
