% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircPalette.R
\name{plotCircPalette}
\alias{plotCircPalette}
\title{plotCircPalette}
\usage{
plotCircPalette(col, num)
}
\arguments{
\item{col}{a single column data frame labeled HEX in which the hexadecimal color codes have been stored}

\item{num}{the number of color codes inputted}
}
\value{
a color palette in color wheel form
}
\description{
plotCircPalette
}
