% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPalette.R
\name{plotPalette}
\alias{plotPalette}
\title{plotPalette}
\usage{
plotPalette(col, num)
}
\arguments{
\item{col}{a single column data frame labeled HEX in which the hexadecimal color codes have been stored}

\item{num}{the number of color codes inputted}
}
\value{
a color palette in stacked barplot form
}
\description{
converts input of hexadecimal color codes into a simple stack of all the colors
}
